cleanup: move outpoint to basetypes package
